+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/TransactionReceiptResult.java
@@ -82,7 +82,8 @@ public abstract class TransactionReceiptResult {
             receiptWithMetadata.getBlockNumber(),
             txn.getHash(),
             receiptWithMetadata.getBlockHash(),
-            receiptWithMetadata.getTransactionIndex());
+            receiptWithMetadata.getTransactionIndex(),
+            receiptWithMetadata.getLogIndexStart());
     this.logsBloom = receipt.getBloomFilter().toString();
     this.to = txn.getTo().map(Bytes::toHexString).orElse(null);
     this.transactionHash = txn.getHash().toString();
@@ -161,14 +162,15 @@ public abstract class TransactionReceiptResult {
       final long blockNumber,
       final Hash transactionHash,
       final Hash blockHash,
-      final int transactionIndex) {
+      final int transactionIndex,
+      final int logIndexStart) {
     final List<TransactionReceiptLogResult> logResults = new ArrayList<>(logs.size());
 
     for (int i = 0; i < logs.size(); i++) {
       final Log log = logs.get(i);
       logResults.add(
           new TransactionReceiptLogResult(
-              log, blockNumber, transactionHash, blockHash, transactionIndex, i));
+              log, blockNumber, transactionHash, blockHash, transactionIndex, logIndexStart + i));
     }
 
     return logResults;